requirements.txt
flask
pyjwt
requests
selenium
beautifulsoup4
